music_api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pymysql
python-dotenv
requests
pytest
httpx
email-validator